move cache annotation in controller dir
